+++ src/sampler/sampler_cube.rs
@@ -59,7 +59,7 @@ impl<'a, 'b: 'a> FSamplerCube<'a> {
     }
 
     /// Sample the sampler texture at a specific level.
-    pub fn texel_lod(&self, sample_coord: NormalizeCoord2d, face: usize, level: usize) -> [f32; 4] {
+    pub fn texel_lod(&self, sample_coord: NormalizeCoord2d, face: usize, level: f32) -> [f32; 4] {
         let raw = unsafe { bindings::fsampler_cube_texel_lod(&self.ffi, <[f32; 2]>::from(sample_coord).as_ptr(), face, level) };
         raw.content
     }
